ragsearch/requirements.txt
django
pandas
numpy
faiss-cpu
sentence-transformers
rank_bm25
openai
scikit-learn
tqdm
